crypto_playground/requirements.txt
alpaca-py>=0.8.0
pandas>=1.5.0
numpy>=1.21.0
ta-lib>=0.4.0
python-dotenv>=0.19.0
websockets>=10.0
aiohttp>=3.8.0
asyncio>=3.4.3
